+++ eruption-macro/src/main.rs
@@ -194,6 +194,7 @@ fn print_header() {
     println!();
 }
 
+/*
 #[cfg(debug_assertions)]
 mod thread_util {
     use crate::Result;
@@ -226,6 +227,7 @@ mod thread_util {
         Ok(())
     }
 }
+*/
 
 pub async fn async_main() -> std::result::Result<(), eyre::Error> {
     cfg_if::cfg_if! {
@@ -255,9 +257,9 @@ pub async fn async_main() -> std::result::Result<(), eyre::Error> {
     }
 
     // start the thread deadlock detector
-    #[cfg(debug_assertions)]
-    thread_util::deadlock_detector()
-        .unwrap_or_else(|e| error!("Could not spawn deadlock detector thread: {}", e));
+    // #[cfg(debug_assertions)]
+    // thread_util::deadlock_detector()
+    //     .unwrap_or_else(|e| error!("Could not spawn deadlock detector thread: {}", e));
 
     // register ctrl-c handler
     let (ctrl_c_tx, _ctrl_c_rx) = unbounded();