+++ eruption/src/scripts/hwaccel.lua
@@ -1,0 +1,49 @@
+-- This file is part of Eruption.
+--
+-- Eruption is free software: you can redistribute it and/or modify
+-- it under the terms of the GNU General Public License as published by
+-- the Free Software Foundation, either version 3 of the License, or
+-- (at your option) any later version.
+--
+-- Eruption is distributed in the hope that it will be useful,
+-- but WITHOUT ANY WARRANTY without even the implied warranty of
+-- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+-- GNU General Public License for more details.
+--
+-- You should have received a copy of the GNU General Public License
+-- along with Eruption.  If not, see <http://www.gnu.org/licenses/>.
+--
+-- Copyright (c) 2019-2022, The Eruption Development Team
+--
+require "declarations"
+require "debug"
+
+-- global state variables --
+color_map = {}
+ticks = 0
+max_effect_ttl = target_fps * 8
+effect_ttl = max_effect_ttl
+
+-- event handler functions --
+function on_startup(config)
+    local accel_info = query_hw_accel_info()
+    debug("HW accel status: " .. stringify(accel_info))
+
+    for i = 1, canvas_size do color_map[i] = 0x00000000 end
+end
+
+function on_key_down(key_index) effect_ttl = max_effect_ttl end
+
+function on_key_up(key_index) effect_ttl = max_effect_ttl end
+
+function on_tick(delta)
+    ticks = ticks + delta
+
+    if effect_ttl <= 0 then return end
+
+    effect_ttl = effect_ttl - 1
+
+    -- color_map = color_map_from_render_surface()
+
+    submit_color_map(color_map)
+end